+++ packages/app/src/initialization/plugin-host/register.ts
