+++ packages/app/src/initialization/queryClient.ts
@@ -1,0 +1,6 @@
+import { queryClient } from '@masknet/shared-base-ui'
+import { broadcastQueryClient } from '@tanstack/query-broadcast-client-experimental'
+
+export function initQueryClient() {
+    broadcastQueryClient({ queryClient })
+}
