+++ packages/app/src/initialization/rpc.ts
