+++ packages/flags/src/flags/buildInfo.ts
@@ -1,4 +1,4 @@
-import { Environment, getEnvironment } from '@dimensiondev/holoflows-kit'
+import { Environment, isEnvironment } from '@dimensiondev/holoflows-kit'
 import { defer } from '@masknet/kit'
 
 export interface BuildInfoFile {
@@ -14,8 +14,7 @@ export interface BuildInfoFile {
 
 export async function getBuildInfo(): Promise<BuildInfoFile> {
     try {
-        // eslint-disable-next-line no-bitwise
-        const hasBrowserAPI = getEnvironment() & Environment.HasBrowserAPI
+        const hasBrowserAPI = isEnvironment(Environment.HasBrowserAPI)
         const b = (globalThis as any).browser
         const manifestVersion = hasBrowserAPI ? b.runtime.getManifest().version : undefined
         const response = await fetch(hasBrowserAPI ? b.runtime.getURL('/build-info.json') : '/build-info.json')
@@ -33,11 +32,9 @@ export let env: BuildInfoFile
 const [_promise, resolve] = defer<void>()
 export const buildInfoReadyPromise = _promise
 export async function setupBuildInfo(): Promise<void> {
-    if (env) return
     return setupBuildInfoManually(await getBuildInfo())
 }
 export function setupBuildInfoManually(_env: BuildInfoFile) {
-    if (env) return
     resolve()
     env = _env
 }
